Endre mentorAntallTimer til Double (#1021)
* Endre mentorAntallTimer til Double

og decimal i postgres

* Rename database migrering fil

Sånn at den blir neste i versjonsnummerrekka

+++ src/main/java/no/nav/tag/tiltaksgjennomforing/avtale/EndreAvtale.java
@@ -63,7 +63,7 @@ public class EndreAvtale {
     private String mentorFornavn;
     private String mentorEtternavn;
     private String mentorOppgaver;
-    private Integer mentorAntallTimer;
+    private Double mentorAntallTimer;
     private String mentorTlf;
     private Integer mentorTimelonn;
